+++ core/control-plane/control-plane-aggregate-services/build.gradle.kts
@@ -23,7 +23,7 @@ dependencies {
     implementation(project(":spi:common:boot-spi"))
     implementation(project(":spi:common:transaction-spi"))
     implementation(project(":spi:control-plane:asset-spi"))
-    implementation(project(":spi:control-plane:secret-spi"))
+    implementation(project(":spi:control-plane:secrets-spi"))
     implementation(project(":spi:control-plane:transfer-data-plane-spi"))
 
     implementation(libs.opentelemetry.instrumentation.annotations)
