+++ core/control-plane/control-plane-aggregate-services/src/main/java/org/eclipse/edc/connector/controlplane/services/secret/SecretServiceImpl.java
@@ -16,12 +16,9 @@ package org.eclipse.edc.connector.controlplane.services.secret;
 
 import org.eclipse.edc.connector.secret.spi.observe.SecretObservable;
 import org.eclipse.edc.connector.spi.service.SecretService;
-import org.eclipse.edc.spi.query.QuerySpec;
 import org.eclipse.edc.spi.result.ServiceResult;
 import org.eclipse.edc.spi.security.Vault;
 import org.eclipse.edc.spi.types.domain.secret.Secret;
-
-import java.util.List;
 
 import static java.util.Optional.ofNullable;
 import static org.eclipse.edc.spi.result.ServiceResult.badRequest;
@@ -46,11 +43,6 @@ public class SecretServiceImpl implements SecretService {
                         .id(secretId)
                         .build())
                 .orElse(null);
-    }
-
-    @Override
-    public ServiceResult<List<Secret>> search(QuerySpec query) {
-        throw new UnsupportedOperationException("Query operation is not supported for secrets");
     }
 
     @Override
