--- extensions/control-plane/api/management-api/secret-api/src/main/java/org/eclipse/edc/connector/api/management/secret/SecretApiExtension.java
@@ -1,63 +1,0 @@
-/*
- *  Copyright (c) 2024 Amadeus
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Amadeus - Initial API and Implementation
- *
- */
-
-package org.eclipse.edc.connector.api.management.secret;
-
-import org.eclipse.edc.connector.api.management.configuration.ManagementApiConfiguration;
-import org.eclipse.edc.connector.api.management.secret.validation.SecretValidator;
-import org.eclipse.edc.connector.spi.service.SecretService;
-import org.eclipse.edc.runtime.metamodel.annotation.Extension;
-import org.eclipse.edc.runtime.metamodel.annotation.Inject;
-import org.eclipse.edc.spi.system.ServiceExtension;
-import org.eclipse.edc.spi.system.ServiceExtensionContext;
-import org.eclipse.edc.transform.spi.TypeTransformerRegistry;
-import org.eclipse.edc.validator.spi.JsonObjectValidatorRegistry;
-import org.eclipse.edc.web.spi.WebService;
-
-import static org.eclipse.edc.spi.types.domain.secret.Secret.EDC_SECRET_TYPE;
-
-@Extension(value = SecretApiExtension.NAME)
-public class SecretApiExtension implements ServiceExtension {
-
-    public static final String NAME = "Management API: Secret";
-
-    @Inject
-    private WebService webService;
-
-    @Inject
-    private ManagementApiConfiguration config;
-
-    @Inject
-    private TypeTransformerRegistry transformerRegistry;
-
-    @Inject
-    private SecretService secretService;
-
-    @Inject
-    private JsonObjectValidatorRegistry validator;
-
-    @Override
-    public String name() {
-        return NAME;
-    }
-
-    @Override
-    public void initialize(ServiceExtensionContext context) {
-        var monitor = context.getMonitor();
-
-        validator.register(EDC_SECRET_TYPE, SecretValidator.instance());
-
-        webService.registerResource(config.getContextAlias(), new SecretApiController(secretService, transformerRegistry, monitor, validator));
-    }
-}
