--- extensions/control-plane/api/management-api/secret-api/src/main/java/org/eclipse/edc/connector/api/management/secret/validation/SecretValidator.java
@@ -1,36 +1,0 @@
-/*
- *  Copyright (c) 2024 Amadeus
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Amadeus - Initial API and Implementation
- *
- */
-
-package org.eclipse.edc.connector.api.management.secret.validation;
-
-import jakarta.json.JsonObject;
-import org.eclipse.edc.validator.jsonobject.JsonObjectValidator;
-import org.eclipse.edc.validator.jsonobject.validators.MandatoryValue;
-import org.eclipse.edc.validator.jsonobject.validators.OptionalIdNotBlank;
-import org.eclipse.edc.validator.spi.Validator;
-
-import static org.eclipse.edc.spi.types.domain.secret.Secret.EDC_SECRET_VALUE;
-
-/**
- * Contains the SecretEntryDto validator definition
- */
-public class SecretValidator {
-
-    public static Validator<JsonObject> instance() {
-        return JsonObjectValidator.newValidator()
-                .verifyId(OptionalIdNotBlank::new)
-                .verify(EDC_SECRET_VALUE, MandatoryValue::new)
-                .build();
-    }
-}
