--- extensions/control-plane/api/management-api/secret-api/src/test/java/org/eclipse/edc/connector/api/management/secret/SecretApiExtensionTest.java
@@ -1,57 +1,0 @@
-/*
- *  Copyright (c) 2024 Amadeus
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Amadeus - Initial API and Implementation
- *
- */
-
-package org.eclipse.edc.connector.api.management.secret;
-
-import org.eclipse.edc.junit.extensions.DependencyInjectionExtension;
-import org.eclipse.edc.spi.system.ServiceExtensionContext;
-import org.eclipse.edc.validator.spi.JsonObjectValidatorRegistry;
-import org.eclipse.edc.web.spi.WebService;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-
-import static org.eclipse.edc.spi.types.domain.secret.Secret.EDC_SECRET_TYPE;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.ArgumentMatchers.isA;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
-
-@ExtendWith(DependencyInjectionExtension.class)
-class SecretApiExtensionTest {
-
-    private final JsonObjectValidatorRegistry validatorRegistry = mock();
-    private final WebService webService = mock();
-
-    @BeforeEach
-    void setUp(ServiceExtensionContext context) {
-        context.registerService(JsonObjectValidatorRegistry.class, validatorRegistry);
-        context.registerService(WebService.class, webService);
-    }
-
-    @Test
-    void initialize_shouldRegisterControllers(SecretApiExtension extension, ServiceExtensionContext context) {
-        extension.initialize(context);
-
-        verify(webService).registerResource(any(), isA(SecretApiController.class));
-    }
-
-    @Test
-    void initialize_shouldRegisterValidators(SecretApiExtension extension, ServiceExtensionContext context) {
-        extension.initialize(context);
-
-        verify(validatorRegistry).register(eq(EDC_SECRET_TYPE), any());
-    }
-}
