+++ extensions/control-plane/api/management-api/secrets-api/build.gradle.kts
@@ -1,0 +1,46 @@
+/*
+ * Copyright (c) 2024 Amadeus
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Apache License, Version 2.0 which is available at
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * Contributors:
+ *    Amadeus - Initial API and Implementation
+ *
+ */
+
+
+plugins {
+    `java-library`
+    id("io.swagger.core.v3.swagger-gradle-plugin")
+}
+
+dependencies {
+    api(project(":spi:common:validator-spi"))
+    api(project(":spi:control-plane:control-plane-spi"))
+    implementation(project(":extensions:common:api:api-core"))
+    implementation(project(":extensions:common:api:management-api-configuration"))
+    implementation(project(":core:common:lib:validator-lib"))
+    implementation(project(":core:common:lib:transform-lib"))
+
+    implementation(libs.jakarta.rsApi)
+
+    testImplementation(project(":core:control-plane:control-plane-core"))
+    testImplementation(project(":core:data-plane-selector:data-plane-selector-core"))
+    testImplementation(project(":extensions:common:http"))
+    testImplementation(project(":core:common:junit"))
+    testImplementation(testFixtures(project(":extensions:common:http:jersey-core")))
+    testImplementation(libs.restAssured)
+    testImplementation(libs.awaitility)
+}
+
+edcBuild {
+    swagger {
+        apiGroup.set("management-api")
+    }
+}
+
+
