+++ settings.gradle.kts
@@ -160,7 +160,7 @@ include(":extensions:control-plane:api:control-plane-api")
 include(":extensions:control-plane:api:control-plane-api-client")
 include(":extensions:control-plane:api:management-api")
 include(":extensions:control-plane:api:management-api:asset-api")
-include(":extensions:control-plane:api:management-api:secret-api")
+include(":extensions:control-plane:api:management-api:secrets-api")
 include(":extensions:control-plane:api:management-api:catalog-api")
 include(":extensions:control-plane:api:management-api:contract-agreement-api")
 include(":extensions:control-plane:api:management-api:contract-definition-api")
@@ -250,7 +250,7 @@ include(":spi:control-plane:control-plane-spi")
 include(":spi:control-plane:transfer-data-plane-spi")
 include(":spi:control-plane:policy-spi")
 include(":spi:control-plane:transfer-spi")
-include(":spi:control-plane:secret-spi")
+include(":spi:control-plane:secrets-spi")
 include(":spi:control-plane:control-plane-api-client-spi")
 
 include(":spi:data-plane:data-plane-spi")
