+++ spi/common/core-spi/src/test/java/org/eclipse/edc/spi/types/domain/secret/SecretTest.java
@@ -14,38 +14,15 @@
 
 package org.eclipse.edc.spi.types.domain.secret;
 
-import org.eclipse.edc.json.JacksonTypeManager;
-import org.eclipse.edc.spi.types.TypeManager;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatNullPointerException;
 
 class SecretTest {
-
-    private TypeManager typeManager;
-
-    @BeforeEach
-    void setUp() {
-        typeManager = new JacksonTypeManager();
-    }
-
-    @Test
-    void verifySerDeser() {
-        var secret = Secret.Builder.newInstance().id("abcd123")
-                .value("valuetest")
-                .build();
-
-        var serialized = typeManager.writeValueAsString(secret);
-        var deserialized = typeManager.readValue(serialized, Secret.class);
-
-        assertThat(deserialized).usingRecursiveComparison().isEqualTo(secret);
-    }
-
+    
     @Test
     void verifyExceptionWhenValueMissing() {
         assertThatNullPointerException().isThrownBy(() -> Secret.Builder.newInstance().id("abcd123").build())
-                .withMessage("`value` is missing");
+                .withMessage("'value' is missing");
     }
 }
