+++ spi/control-plane/secrets-spi/src/main/java/org/eclipse/edc/connector/secret/spi/event/SecretDeleted.java
