+++ spi/control-plane/secrets-spi/src/main/java/org/eclipse/edc/connector/secret/spi/observe/SecretObservableImpl.java
@@ -1,0 +1,23 @@
+/*
+ *  Copyright (c) 2024 Amadeus
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Amadeus - Initial API and Implementation
+ *
+ */
+
+package org.eclipse.edc.connector.secret.spi.observe;
+
+import org.eclipse.edc.spi.observe.ObservableImpl;
+
+/**
+ * Observes Secret lifecycle state changes.
+ */
+public class SecretObservableImpl extends ObservableImpl<SecretListener> implements SecretObservable {
+}
