+++ spi/control-plane/secrets-spi/src/test/java/org/eclipse/edc/connector/contract/spi/event/SecretEventTest.java
