+++ system-tests/management-api/management-api-test-runtime/build.gradle.kts
@@ -24,6 +24,7 @@ dependencies {
     implementation(project(":extensions:common:iam:iam-mock"))
     implementation(project(":extensions:common:json-ld"))
     implementation(project(":extensions:control-plane:api:management-api"))
+    implementation(project(":extensions:control-plane:api:management-api:secrets-api"))
     implementation(project(":extensions:control-plane:transfer:transfer-data-plane"))
     implementation(project(":extensions:data-plane:data-plane-client"))
